ecommerce/requirement.txt
asgiref==3.8.1
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
